+++ tests/unit_tests/recipes/test_base_recipe.py
@@ -19,8 +19,8 @@ import pytest
 import torch
 import torch.nn as nn
 
-import nemo_automodel.components.training.base_recipe as base_recipe
-from nemo_automodel.components.training.base_recipe import BaseRecipe, _find_latest_checkpoint
+import nemo_automodel.recipes.base_recipe as base_recipe
+from nemo_automodel.recipes.base_recipe import BaseRecipe, _find_latest_checkpoint
 
 try:
     import expecttest
